backend/requirements.txt
tinytuya>=1.13
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
alembic>=1.13
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
httpx>=0.27
